handle ref to clauses in ScenarioVariant, refactor registration
* Update iter_fluent_refs() to also handle HoldsExpr in ScenarioVariant.
* Use classes instead of string for matching model types.
* Split language and generator defs to separate modules, keep only
  LanguageDesc & GeneratorDesc calls in registration.
